chore(preview): split tree_sitter.rs into tree_sitter_grammars.rs and tree_sitter_highlight.rs
